refact unit test
