+++ .github/workflows/pr-test-xpu.yml
@@ -67,6 +67,7 @@ jobs:
               python3 bench_moe_sum_reduce.py 2>&1 | tee moe_sum_reduce.log \
               python3 bench_moe_fused_gate.py 2>&1 | tee moe_fused_gate.py.log \
               python3 bench_merge_states_v2.py 2>&1 | tee merge_states.py.log \
+              python3 bench_mrope.py 2>&1 | tee mrope.py.log \
               python3 bench_swiglu_alpha_limit.py 2>&1 | tee swiglu_alpha_limit.py.log \
               python3 bench_fused_qk_norm_rope.py 2>&1 | tee fused_qk_norm_rope.py.log \
               python3 bench_fused_qk_rope_with_cache.py 2>&1 | tee bench_fused_qk_rope_with_cache.py.log \
